requirements.txt
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1696634205638/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1708355285029/work
archspec @ file:///croot/archspec_1709217642129/work
boltons @ file:///C:/b/abs_707eo7c09t/croot/boltons_1677628723117/work
Bottleneck @ file:///C:/b/abs_f05kqh7yvj/croot/bottleneck_1707864273291/work
Brotli @ file:///C:/Windows/Temp/abs_63l7912z0e/croots/recipe/brotli-split_1659616056886/work
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///D:/bld/click_1692312014553/work
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
conda @ file:///C:/b/abs_f3egtzumcq/croot/conda_1715635886039/work
conda-content-trust @ file:///C:/b/abs_e3bcpyv7sw/croot/conda-content-trust_1693490654398/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///C:/b/abs_b9wp3lr1gn/croot/conda-package-handling_1691008700066/work
conda_package_streaming @ file:///C:/b/abs_6c28n38aaj/croot/conda-package-streaming_1690988019210/work
cryptography @ file:///C:/b/abs_f5n93r0tun/croot/cryptography_1710350404202/work
distro @ file:///C:/b/abs_a3uni_yez3/croot/distro_1701455052240/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1691598144738/work
email_validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1709024422129/work
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1715196982285/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/fastapi-cli_1716205772095/work
frozendict @ file:///D:/bld/frozendict_1715092910909/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1711596990900/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1708530890843/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
jsonpointer==2.1
language-tool-python @ file:///home/conda/feedstock_root/build_artifacts/language_tool_python_1610478442101/work
libmambapy @ file:///C:/b/abs_7dmjutgtwb/croot/mamba-split_1712091963973/work/libmambapy
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe @ file:///D:/bld/markupsafe_1706900063757/work
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
menuinst @ file:///C:/b/abs_099kybla52/croot/menuinst_1706732987063/work
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
nltk @ file:///home/conda/feedstock_root/build_artifacts/nltk_1672696305909/work
numexpr @ file:///C:/b/abs_5fucrty5dc/croot/numexpr_1696515448831/work
openpyxl @ file:///C:/b/abs_764brjadj1/croot/openpyxl_1714158895873/work
packaging @ file:///C:/b/abs_cc1h2xfosn/croot/packaging_1710807447479/work
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
pluggy @ file:///C:/ci/pluggy_1648024580010/work
pycosat @ file:///C:/b/abs_31zywn1be3/croot/pycosat_1696537126223/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1713905669586/work
pydantic_core @ file:///D:/bld/pydantic-core_1713862047178/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyMySQL @ file:///C:/ci/pymysql_1610464861530/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///C:/b/abs_a6ui0hquwx/croot/python-dateutil_1715108843988/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/python-multipart_1707760088566/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
regex @ file:///D:/bld/regex_1715828569757/work
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich-split_1709150387247/work/dist
ruamel.yaml @ file:///C:/b/abs_30ee5qbthd/croot/ruamel.yaml_1666304562000/work
ruamel.yaml.clib @ file:///C:/b/abs_aarblxbilo/croot/ruamel.yaml.clib_1666302270884/work
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1698144360966/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1708952932303/work
soupsieve @ file:///C:/b/abs_bbsvy9t4pl/croot/soupsieve_1696347611357/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1709667058396/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
typer==0.12.3
typer-slim==0.12.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
ujson @ file:///D:/bld/ujson_1715783264372/work
urllib3 @ file:///C:/b/abs_4etpfrkumr/croot/urllib3_1707770616184/work
uvicorn @ file:///D:/bld/uvicorn-split_1710932160495/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
zstandard==0.19.0
